Replace float64 with Decimal type for accurate currency conversions

+++ deps.go
@@ -12,11 +12,11 @@ type Cache[T any] interface {
 }
 
 type Source interface {
-	Rate(from, to string) (float64, error)
+	Rate(from, to string) (Decimal, error)
 	Currencies() (map[string]string, error)
 }
 
 type Converter interface {
-	Conv(amount float64, from, to string) (float64, error)
+	Conv(amount Decimal, from, to string) (Decimal, error)
 	Currencies() map[string]string
 }
